pyproject.toml
[tool.poetry]
name = "missue-tracker-linebot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6"
flask = "^1.1"
line-bot-sdk = "^1.15"
flask-sqlalchemy = "^2.4"
pymysql = "^0.9.3"

[tool.poetry.dev-dependencies]
pylint = "^2.4"
autopep8 = "^1.4"
rope = "^0.14.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
